api/requirements.txt
Django==6.0.4
mysqlclient==2.2.8
python-dotenv==1.1.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
